Add typescript to PileupTrack model (#919)
